backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.0
python-barcode==0.15.1
Pillow==10.1.0
twilio==8.10.0
